YouTubeMDBot/requirements.txt
mutagen
isodate
google-api-python-client
musicbrainzngs
ujson
youtube_dl
pyacoustid
python-telegram-bot
